# docs/swagger.yaml
basePath: /api/v1
definitions:
  admin.UpdateGameScoreResponse:
    properties:
      audit_log_id:
        type: string
      away_score:
        type: integer
      game_id:
        type: string
      home_score:
        type: integer
      updated_at:
        type: string
    type: object
  admin.UpdateSpiritScoreResponse:
    properties:
      attitude:
        type: integer
      audit_log_id:
        type: string
      communication:
        type: integer
      fair_mindedness:
        type: integer
      fouls:
        type: integer
      rules_knowledge:
        type: integer
      spirit_score_id:
        type: string
      total_score:
        type: integer
      updated_at:
        type: string
    type: object
  analytics.Dashboard:
    properties:
      dashboard_title:
        type: string
      dashboard_uuid:
        type: string
      id:
        type: integer
      published:
        type: boolean
      slug:
        type: string
      url:
        type: string
    type: object
  analytics.EventStatistics:
    properties:
      average_spirit_score:
        type: number
      completed_games:
        type: integer
      event_id:
        type: string
      in_progress_games:
        type: integer
      scheduled_games:
        type: integer
      top_scorer:
        $ref: '#/definitions/analytics.PlayerStat'
      total_games:
        type: integer
      total_players:
        type: integer
      total_teams:
        type: integer
    type: object
  analytics.GenerateEmbedTokenResponse:
    properties:
      dashboard_uuid:
        type: string
      expires_at:
        type: string
      token:
        type: string
    type: object
  analytics.ListDashboardsResponse:
    properties:
      dashboards:
        items:
          $ref: '#/definitions/analytics.Dashboard'
        type: array
      total:
        type: integer
    type: object
  analytics.NaturalLanguageQueryResponse:
    properties:
      confidence:
        type: number
      explanation:
        type: string
      question:
        type: string
      results:
        items:
          additionalProperties: true
          type: object
        type: array
      sql:
        type: string
      warning:
        type: string
    type: object
  analytics.PlayerStat:
    properties:
      assists:
        type: integer
      games_played:
        type: integer
      goals:
        type: integer
      player_id:
        type: string
      player_name:
        type: string
      team_id:
        type: string
      team_name:
        type: string
    type: object
...
  handlers.AuditLogDTO:
    properties:
      action:
        example: update
        type: string
      changes:
        additionalProperties: true
        type: object
      created_at:
        example: "2026-02-04T12:00:00Z"
        type: string
      entity_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      entity_type:
        example: game
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      ip_address:
        example: 192.168.1.1
        type: string
      reason:
        example: Score correction
        type: string
      user_agent:
        example: Mozilla/5.0
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440002
        type: string
      username:
        example: admin@example.com
        type: string
    type: object
...
  handlers.ErrorResponse:
    description: Error response structure
    properties:
      error:
        example: error message
        type: string
    type: object
...
  handlers.GenerateEmbedTokenRequestDTO:
    properties:
      event_id:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      team_ids:
        items:
          type: string
        type: array
      user_id:
        type: string
      username:
        type: string
    required:
    - first_name
    - last_name
    - user_id
    - username
    type: object
...
  handlers.NaturalLanguageQueryRequestDTO:
    properties:
      context:
        type: string
      event_id:
        type: string
      question:
        type: string
      user_id:
        type: string
    required:
    - question
    - user_id
    type: object
...
  handlers.UpdateGameScoreRequestDTO:
    properties:
      away_score:
        minimum: 0
        type: integer
      home_score:
        minimum: 0
        type: integer
      reason:
        minLength: 10
        type: string
    required:
    - reason
    type: object
  handlers.UpdateSpiritScoreRequestDTO:
    properties:
      attitude:
        maximum: 4
        minimum: 0
        type: integer
      communication:
        maximum: 4
        minimum: 0
        type: integer
      fair_mindedness:
        maximum: 4
        minimum: 0
        type: integer
      fouls:
        maximum: 4
        minimum: 0
        type: integer
      reason:
        minLength: 10
        type: string
      rules_knowledge:
        maximum: 4
        minimum: 0
        type: integer
    required:
    - reason
    type: object
...
host: localhost:4000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for DigiGameStats reimplementation.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: DigiGameStats API
  version: "1.0"
paths:
  /admin/games/{id}/audit:
    get:
      description: Retrieves all administrative changes made to a game
      parameters:
      - description: Game ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.AuditLogDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get game audit history (Admin only)
      tags:
      - admin
  /admin/games/{id}/score:
    put:
      consumes:
      - application/json
      description: Updates a game's score with audit trail
      parameters:
      - description: Score update request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateGameScoreRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.UpdateGameScoreResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update game score (Admin only)
      tags:
      - admin
  /admin/spirit-scores/{id}:
    put:
      consumes:
      - application/json
      description: Updates a spirit score with audit trail
      parameters:
      - description: Spirit Score ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Spirit score update request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateSpiritScoreRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.UpdateSpiritScoreResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update spirit score (Admin only)
      tags:
      - admin
  /analytics/dashboards:
    get:
      consumes:
      - application/json
      description: Retrieves all available analytics dashboards
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/analytics.ListDashboardsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all Superset dashboards
      tags:
      - analytics
  /analytics/dashboards/{dashboard_uuid}:
    get:
      consumes:
      - application/json
      description: Retrieves a specific dashboard's metadata
      parameters:
      - description: Dashboard UUID
        format: uuid
        in: path
        name: dashboard_uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/analytics.Dashboard'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get dashboard by UUID
      tags:
      - analytics
  /analytics/embed-token/{dashboard_uuid}:
    post:
      consumes:
      - application/json
      description: Creates a guest token for embedding a Superset dashboard with RLS
      parameters:
      - description: Dashboard UUID
        format: uuid
        in: path
        name: dashboard_uuid
        required: true
        type: string
      - description: Embed token parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.GenerateEmbedTokenRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/analytics.GenerateEmbedTokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Generate embed token for dashboard
      tags:
      - analytics
  /analytics/events/{event_id}/statistics:
    get:
      consumes:
      - application/json
      description: Retrieves comprehensive analytics for an event
      parameters:
      - description: Event UUID
        format: uuid
        in: path
        name: event_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/analytics.EventStatistics'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get event statistics
      tags:
      - analytics
...
  /analytics/query:
    post:
      consumes:
      - application/json
      description: Converts natural language to SQL using Ollama LLM and executes
        the query
      parameters:
      - description: Natural language question
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.NaturalLanguageQueryRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/analytics.NaturalLanguageQueryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Process natural language query
      tags:
      - analytics
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"